[#454] Move OT::Email to OT::App::Mail::Views
This brings it into parity with the web, api, and colonel views.
