+++ config.ru
@@ -16,9 +16,7 @@ $LOAD_PATH.unshift(File.join(ENV.fetch('APP_ROOT', nil), 'lib'))
 $LOAD_PATH.unshift(File.join(ENV.fetch('APP_ROOT', nil), 'app'))
 
 require 'otto'
-require 'onetime/app/web'
-require 'onetime/app/api'
-require 'onetime/app/colonel'
+require 'onetime/app'
 
 PUBLIC_DIR = "#{ENV.fetch('APP_ROOT', nil)}/public/web".freeze
 APP_DIR = "#{ENV.fetch('APP_ROOT', nil)}/lib/onetime/app".freeze
