+++ lib/onetime/app.rb
@@ -1,0 +1,4 @@
+require_relative 'app/web'
+require_relative 'app/api'
+require_relative 'app/colonel'
+require_relative 'app/mail'
