+++ lib/onetime/app/mail.rb
@@ -1,0 +1,4 @@
+
+require_relative 'mail/base'
+require_relative 'mail/smtp_mailer'
+require_relative 'mail/sendgrid_mailer'
